Rename test.cc -> easy6502, update Makefile
This test isn't an automated test, it involves the user checking the
output of the register and memory dumps to make sure things are as they
should be. It's been renamed and easy6502 credited in the source to
reflect where it comes from and what it's for.

--HG--
rename : src/test.cc => src/easy6502.cc